Update code for Generator
