+++ seq2seq/seq2seq_model.py
@@ -3,36 +3,9 @@ import math
 import torch
 from torch import nn, matmul, softmax, Tensor
 
+from seq2seq.model_utils import TokenEmbedding, PositionalEncoding
 from seq2seq.transformer_decoder import TransformerDecoder
 from seq2seq.transformer_encoder import TransformerEncoder
-
-
-class TokenEmbedding(nn.Module):
-    def __init__(self, vocab_size: int, emb_size):
-        super(TokenEmbedding, self).__init__()
-        self.embedding = nn.Embedding(vocab_size, emb_size)
-        self.emb_size = emb_size
-
-    def forward(self, tokens: Tensor):
-        return self.embedding(tokens.long()) * math.sqrt(self.emb_size)
-
-
-class PositionalEncoding(nn.Module):
-    def __init__(self, emb_size: int, dropout, maxlen: int = 5000):
-        super(PositionalEncoding, self).__init__()
-        den = torch.exp(- torch.arange(0, emb_size, 2) * math.log(10000) / emb_size)
-        pos = torch.arange(0, maxlen).reshape(maxlen, 1)
-        pos_embedding = torch.zeros((maxlen, emb_size))
-        pos_embedding[:, 0::2] = torch.sin(pos * den)
-        pos_embedding[:, 1::2] = torch.cos(pos * den)
-        pos_embedding = pos_embedding.unsqueeze(-2)
-
-        self.dropout = nn.Dropout(dropout)
-        self.register_buffer('pos_embedding', pos_embedding)
-
-    def forward(self, token_embedding: Tensor):
-        return self.dropout(token_embedding +
-                            self.pos_embedding[:token_embedding.size(0), :])
 
 
 class MultiheadedAttention(nn.Module):
